Fixed few typos.
Hi Thomas and team, this was my first attempt to proofread articles here. Last 5/26/2020, I chose "Commit directly to the master branch."
Revisited this today 5/29/2020 and chose the correct option: "Create a new branch for this commit and start a pull request."

Btw, as for the Pixabay stock photo link at the onset, I left it as it is.

My apologies and please be guided accordingly. Thank you eh.